requirements.txt
React
cvzone
numpy
CORS
Flask
Flask-CORS
mysql-connector-python
opencv-python